Update lib, fix get_checked_items
