+++ frappe/public/build.json
@@ -333,7 +333,7 @@
 	"css/report.min.css": [
 		"public/css/report.css",
 		"public/css/tree_grid.css",
-		"public/css/regrid.css",
+		"public/css/frappe-datatable.css",
 
 		"public/js/lib/slickgrid/slick.grid.css",
 		"public/js/lib/slickgrid/slick-default-theme.css",
@@ -341,7 +341,7 @@
 	],
 	"js/report.min.js": [
 		"public/js/lib/clusterize.min.js",
-		"public/js/lib/regrid.js",
+		"public/js/lib/frappe-datatable.js",
 		"public/js/frappe/views/reports/reportview.js",
 		"public/js/frappe/views/reports/reportview_footer.html",
 		"public/js/frappe/views/reports/query_report.js",
