+++ frappe/public/less/frappe-datatable.less
@@ -30,4 +30,8 @@
 	[data-fieldtype="Color"] .control-input {
 		overflow: hidden;
 	}
+
+	.data-table-col.selected .content {
+		border-color: @blue;
+	}
 }
